% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_speech_politicos.R
\name{as_speech_politicos}
\alias{as_speech_politicos}
\title{Adapta base de datos depliticos para pegar con speech}
\usage{
as_speech_politicos()
}
\value{
data.frame
}
\description{
Adapta base de datos depliticos para pegar con speech
}
\examples{
as_speech_politicos()
}
