% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politico.R
\name{buscar_puy}
\alias{buscar_puy}
\title{Busqueda rapida de politrcios por el nombre}
\usage{
buscar_puy(nombre = character(), contiene = FALSE)
}
\arguments{
\item{nombre}{nombre del politico}

\item{contiene}{valor logico que indica si se desea buscar el valor ingresado en \code{nombre} como parte del apellido o del
nombre de un politico. Por defecto es \code{FALSE}, esto quiere decir que la busqueda que se inicia por defeto es por inicio
de los caracteres ingresado en \code{nombre}.}
}
\value{
data.frame
}
\description{
Permite hacer una busqueda rapida de politico en la base de datos \code{puy::politicos}
}
\details{
La busqueda implementa una expresion regular que busca en la base de politicos a aquellos legisladores y legisladoras
    que comienzan con el valor ingresado en \code{nombre}. Es importante tener presente que los nombres de los politicos en la base
    tienen la siguiente estructura: 'APELLIDO, Nombre'. No es necesario respetar el uso de mayusculas o minusculas pero si
    la secuencia del nombre tal cual esta en la base para obtener un resultado satisfactorio. Por ejemplo, si quiere buscar a Tabare Vazques
    y utiliza \code{buscar_puy(nombre = "Tabare Vazquez")} no va a obtener resultados como si hiciera: \code{buscar_puy(nombre = "Vazquez")} o
    \code{buscar_puy(nombre = "Vazquez, Tabare")}.
}
\examples{
buscar_puy(nombre = "Mujica")
buscar_puy(nombre = "Mujica Cordano")

buscar_puy(nombre = "Larrañaga")
# todos los politicos en los que el primer apellido comienza con la letra 'L'
buscar_puy(nombre = "L")

buscar_puy(nombre = "Lacalle")
buscar_puy(nombre = "Lacalle", contiene = TRUE)

}
