% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\docType{data}
\name{legislaturas}
\alias{legislaturas}
\title{legislaturas}
\format{
data.frame con 46 filas y 4. columnas.
}
\usage{
data(politicos)
}
\description{
Contiene las legislaturas desde 1830 a 2020 y el periodo de tiempo en formato \code{Date}.
}
\details{
\describe{
   \item{legislatura}{Numero de la legislatura}
   \item{inicio}{Fecha de inicio de la legislatura.}
   \item{fin}{Fecha de cierre de la legislatura}
   \item{dias}{Cantidad de dias de la legislatura.}
   \item{periodo}{Anios de la legislatura 'yyyy-yyyy'}
}

La unidad de analisis de esta base de datos es la legislatura.
}
\keyword{datasets}
