% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politico.R
\name{politico}
\alias{politico}
\title{Busqueda rapida de politrcios por el nombre}
\usage{
politico(nombre = character())
}
\arguments{
\item{nombre}{nombre del politico}
}
\value{
data.frame
}
\description{
Permite hacer una busqueda rapida de politico en la base de datos \code{puy::politicos}
}
\details{
La busqueda implementa unaexpresion regular que busca en la base de politicos a aquellos que comienzan con
    el valor ingresado en \code{nombre}. Es importante tener presente que los nombres de los politicos en la base
    tienen la siguiente estructura: APELLIDO, Nombre. No es necesario respetar el uso de mayusculas o minusculas pero si
    la secuencia del nombre tal cual esta en la base para obtener un resultado satisfactorio.
}
\examples{
politico(nombre = "Mujica")
politico(nombre = "Mujica Cordano")
}
