% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politicos.R
\docType{data}
\name{politicos}
\alias{politicos}
\title{politicos}
\format{
data.frame con ... filas y 7 columnas.
}
\usage{
data(politicos)
}
\description{
Contiene el periodo y el cargo que ocupo un politico uruguayo desde 1830 hasta 2020.
}
\details{
\describe{
   \item{politico}{Nombre completo del politico en el formato: 'APELLIDO, Nombre'}
   \item{partido}{Partido al que pertenece el politico en el cargo.}
   \item{fecha_inicio}{Fecha de inicio en el que comienza la gestion en el cargo.}
   \item{fecha_fin}{Fecha de fin en la que finaliza la gestion en el cargo.}
   \item{legislatura}{Legislatura en la que ocupa el cargo. Esta variable es para los legisladores unicamente.}
   \item{cargo}{Tipo de cargo.}
   \item{sexo}{Sexo}
}

...
}
\keyword{datasets}
