% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puy_nombres.R
\name{puy_nombres}
\alias{puy_nombres}
\title{Nombre de los politicos uruguayos}
\usage{
puy_nombres()
}
\value{
data.frame con tres variables:
     \itemize{
            \item{\code{nombre}: nombre de pila del legislador}
            \item{\code{cantidad}: cantidad de veres que aparece ese nombre}
            \item{\code{porcentaje}: porcentaje de veses que aparece ese nombre}
           }
}
\description{
Proporciona la frecuencia de nombres de los politicos uruguayos
}
\examples{
puy_nombres()
}
