% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_speech_politicos.R
\name{as_speech_politicos}
\alias{as_speech_politicos}
\title{Adapta base de datos depliticos para pegar con speech}
\usage{
as_speech_politicos(floor_speech)
}
\arguments{
\item{floor_speech}{diario de sesion en formato data.frame del paquete \code{speech}.}
}
\value{
data.frame
}
\description{
Adapta base de datos depliticos para pegar con speech
}
\examples{
library(speech)
url <- "http://bit.ly/35AUVF4"
text <- speech::speech_build(file = url)
as_speech_politicos(floor_speech = text)
}
