% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_speech_politicos.R
\name{as_speech_politicos}
\alias{as_speech_politicos}
\title{Adapta base de datos depliticos para pegar con speech}
\usage{
as_speech_politicos(speech)
}
\arguments{
\item{speech}{diario de sesion en formato data.frame del paquete \code{speech}.}
}
\value{
data.frame de entrada al que se le agregan 3 variables:
     \itemize{
            \item{\code{legislator2}: Nombre completo del legislador/a}
            \item{\code{party}: partido politico al que pertenece el legislador/a en ese momento}
            \item{\code{indicator}: un indicador de calidad del merge entre el diario de sesion y la base de politicos. Puede asumir los siguientres valores:}
                 \itemize{
                 \item{1 - Match perfecto (nombre del legislador/a, legislatura, caamara y sexo)}
                 \item{2 - Match parcial (igual al anterior pero solo con el primer apellido del legislador en la base de politicos. Sucede que mucahs veces en los diarios de sesion no aparece el legislador con ambos apellidos y en la base de politicos si.}
                 \item{3 - Match parcial (igual que el valor 1 pero cambiando la identificacion de legislatura para atras en una unidad)}
                 \item{4 - Match parcial (igual que el valor 1 pero cambiando la identificacion de legislatura para adelante en una unidad)}
                 \item{5 - Match parcial debil (un solo apellido con la legislatura anterior)}
                 \item{6 - Match parcial debil (un solo apellido con la legislatura posterior)}
                  }

           }
}
\description{
Adapta base de datos depliticos para pegar con speech
}
\examples{
library(speech)
url <- "http://bit.ly/35AUVF4"
text <- speech::speech_build(file = url, compiler = TRUE)
text <- speech::speech_legis_replace(tidy_speech = text, old = "GOI", new = "GONI")
floor_speech <- as_speech_politicos(speech = text)
}
