% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislaturas.R
\docType{data}
\name{legislaturas}
\alias{legislaturas}
\title{legislaturas}
\format{
data.frame con 46 filas y 4. columnas.
}
\usage{
data(politicos)
}
\description{
Contiene las fechas en las que estuvieron abiertas las legislaturas desde 1830 a 2020.
}
\details{
\describe{
   \item{legislatura}{Numero de la legislatura}
   \item{inicio}{Fecha de inicio de la legislatura.}
   \item{fin}{Fecha de cierre de la legislatura}
   \item{dias}{Cantidad de dias que estuvo abierta la legislatura.}
}

La unidad de analisis de esta base de datos es..
}
\keyword{datasets}
