% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politicos.R
\docType{data}
\name{politicos}
\alias{politicos}
\title{politicos}
\format{
data.frame con ... filas y ... columnas.
}
\usage{
data(politicos)
}
\description{
Contiene el periodo y el cargo que ocupo un politico uruguayo desde 1830 hasta 2020.
}
\details{
\describe{
   \item{...}{...}
}

La unidad de analisis de esta base de datos es...
}
\keyword{datasets}
