% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politicos.R
\docType{data}
\name{politicos}
\alias{politicos}
\title{politicos}
\format{
data.frame con ... filas y 7 columnas.
}
\usage{
data(politicos)
}
\description{
Contiene el periodo y el cargo que ocupo un politico uruguayo desde 1830 hasta 2020.
}
\details{
\describe{
   \item{politico}{...}
   \item{partido}{...}
   \item{fecha_inicio}{...}
   \item{fecha_fin}{...}
   \item{legislatura}{...}
   \item{cargo}{...}
   \item{sexo}{...}
}

La unidad de analisis de esta base de datos es...
}
\keyword{datasets}
